requirements.txt
boto3>=1.42.51
botocore>=1.42.51
jmespath>=1.1.0
python-dateutil>=2.9.0.post0
s3transfer>=0.16.0
six>=1.17.0
urllib3>=2.6.3
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.3
aiosignal>=1.4.0
annotated-types>=0.7.0
anyio>=4.12.1
attrs>=25.4.0
cachetools>=7.0.1
certifi>=2026.1.4
cffi>=2.0.0
charset-normalizer>=3.4.4
cryptography>=46.0.5
dataclasses-json>=0.6.7
distro>=1.9.0
filetype>=1.2.0
frozenlist>=1.8.0
google-auth>=2.48.0
google-genai>=1.63.0
greenlet>=3.3.1
grpcio>=1.78.0
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
httpx-sse>=0.4.3
idna>=3.11
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=1.2.10
langchain-classic>=1.0.1
langchain-community>=0.4.1
langchain-core>=1.2.13
langchain-google-genai>=4.2.0
langchain-milvus>=0.3.3
langchain-text-splitters>=1.1.0
langgraph>=1.0.8
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.7
langgraph-sdk>=0.3.6
langsmith>=0.7.3
marshmallow>=3.26.2
multidict>=6.7.1
mypy_extensions>=1.1.0
numpy>=2.4.2
orjson>=3.11.7
ormsgpack>=1.12.2
packaging>=26.0
pandas>=3.0.0
propcache>=0.4.1
protobuf>=6.33.5
pyasn1>=0.6.2
pyasn1_modules>=0.4.2
pycparser>=3.0
pydantic>=2.12.5
pydantic-settings>=2.13.0
pydantic_core>=2.41.5
pymilvus>=2.6.9
pypdf>=6.7.0
python-dotenv>=1.2.1
PyYAML>=6.0.3
requests>=2.32.5
requests-toolbelt>=1.0.0
rsa>=4.9.1
setuptools>=82.0.0
sniffio>=1.3.1
SQLAlchemy>=2.0.46
tenacity>=9.1.4
typing-inspect>=0.9.0
typing-inspection>=0.4.2
typing_extensions>=4.15.0
uuid_utils>=0.14.0
websockets>=15.0.1
xxhash>=3.6.0
yarl>=1.22.0
zstandard>=0.25.0
langchain-openai>=1.1.10
google-cloud-aiplatform>=1.139.0